requirements.txt
markdown
numpy
pandas
Pillow>=9.5.0
pyyaml
tqdm
torch
torchvision
pytorch-lightning
scikit-learn
tensorboard
lightning
pycytominer
scipy==1.8.0